+++ R/fit_SE.combinations.R
@@ -21,7 +21,7 @@
 #' @examples 
 #' fmae_cms <- gDRutils::get_synthetic_data("finalMAE_combo_matrix_small")
 #' 
-#' se1 <- fmae_cms[["matrix"]]
+#' se1 <- fmae_cms[[gDRutils::get_experiment_groups("combination")]]
 #' SummarizedExperiment::assays(se1) <- 
 #'   SummarizedExperiment::assays(se1)["Averaged"]
 #' fit_SE.combinations(se1[1, 1])
@@ -32,7 +32,7 @@
 #' @export
 #'
 fit_SE.combinations <- function(se,
-                                data_type = "matrix",
+                                data_type = gDRutils::get_experiment_groups("combination"),
                                 series_identifiers = NULL,
                                 normalization_types = c("GR", "RV"),
                                 averaged_assay = "Averaged",
@@ -74,10 +74,9 @@ fit_SE.combinations <- function(se,
 
   out <- gDRutils::loop(seq_len(nrow(iterator)), function(row) {
   
-    bliss_excess <- hsa_excess <- metrics <- all_iso_points <- 
-      isobolograms <- smooth_mx <- NULL
-    bliss_score <- hsa_score <- CIScore_50 <- CIScore_80 <- 
-      S4Vectors::DataFrame(normalization_type = normalization_types)
+    metrics <- all_iso_points <- isobolograms <- excess_full <- NULL
+    excess <- scores <- 
+      data.table::data.table(normalization_type = normalization_types)
     x <- iterator[row, ]
     i <- x[["row"]]
     j <- x[["column"]]
@@ -85,23 +84,13 @@ fit_SE.combinations <- function(se,
     avg_combo <- data.table::as.data.table(avg[avg[["row"]] == i & avg[["column"]] == j, ])
    
     if (all(is.na(avg_combo[, -c("row", "column", "normalization_type")]))) { # omit masked values (all NAs)
-      smooth_mx <- hsa_excess <- bliss_excess <- isobolograms <- metrics <- 
-        bliss_score[, c("row_id", "col_id")] <- 
-        hsa_score[, c("row_id", "col_id")] <-
-        CIScore_50[, c("row_id", "col_id")] <- 
-        CIScore_80[, c("row_id", "col_id")] <-
-        all_iso_points <-
+      excess <- isobolograms <- metrics <- scores <- all_iso_points <-
         data.table::data.table(row_id = i, col_id = j)
-      return(list(bliss_excess = bliss_excess,
-           hsa_excess = hsa_excess,
+      return(list(excess = excess,
            metrics = metrics,
            all_iso_points = all_iso_points,
            isobolograms = isobolograms,
-           smooth_mx = smooth_mx,
-           bliss_score = bliss_score,
-           hsa_score = hsa_score,
-           CIScore_50 = CIScore_50,
-           CIScore_80 = CIScore_80))
+           scores = scores))
     }
     
     conc_map <- map_conc_to_standardized_conc(avg_combo[[id]], avg_combo[[id2]])
@@ -242,16 +231,17 @@ fit_SE.combinations <- function(se,
       # store the values from the Averaged assay for reference
       complete_subset$av_values <- complete_subset$x
       # and replace by the Smoothed values
-      complete_subset$x <- rowMeans(mat, na.rm = TRUE)
+      complete_subset$smooth <- rowMeans(mat, na.rm = TRUE)
       
       # just keep the relevant columns and change to the metric name
-      cols <- c(id, id2, "x")
+      cols <- c(id, id2, "smooth")
       av_matrix <- complete_subset[, cols, with = FALSE]
       av_matrix[, "normalization_type" := norm_type]
-      av_matrix[get(id) == 0 & get(id2) == 0, x := 1]
-
+      
+      av_matrix[get(id) == 0 & get(id2) == 0, smooth := 1]
       if (NROW(av_matrix) == 0) {
-        av_matrix <- h_excess <- b_excess <- NULL
+        excess[excess$normalization_type == norm_type,
+               c("smooth", "hsa_excess", "bliss_excess")] <- NA
       } else {
         sa1 <- av_matrix[get(id2) == 0, cols, with = FALSE]
         sa2 <- av_matrix[get(id) == 0, cols, with = FALSE]
@@ -262,26 +252,28 @@ fit_SE.combinations <- function(se,
           av_matrix, 
           series_identifiers = c(id, id2), 
           metric_col = "metric",
-          measured_col = "x"
+          measured_col = "smooth"
         )
-
+        data.table::setnames(h_excess, "x", "hsa_excess")
         bliss <- calculate_Bliss(sa1, id, sa2, id2, norm_type)
-        b_excess <- calculate_excess(
+        bliss_excess <- calculate_excess(
           bliss, 
           av_matrix, 
           series_identifiers = c(id, id2), 
           metric_col = "metric",
-          measured_col = "x"
+          measured_col = "smooth"
         )
+        data.table::setnames(bliss_excess, "x", "bliss_excess")
+        excess <- Reduce(function(x, y) merge(x, y, all = TRUE), list(av_matrix, h_excess, bliss_excess))
       }
       
       # call calculate_Loewe and calculate_isobolograms: 
       # remove rows/columns with less than 2 values
       discard_conc_1 <- names(which(
-        table(av_matrix[!is.na(x) & normalization_type == norm_type, id, with = FALSE]) < 3
+        table(av_matrix[!is.na(smooth) & normalization_type == norm_type, id, with = FALSE]) < 3
       ))
       discard_conc_2 <- names(which(
-        table(av_matrix[!is.na(x) & normalization_type == norm_type, id2, with = FALSE]) < 3
+        table(av_matrix[!is.na(smooth) & normalization_type == norm_type, id2, with = FALSE]) < 3
       ))
       av_matrix_dense <- av_matrix[
         !(av_matrix[[id]] %in% discard_conc_1) & 
@@ -306,79 +298,64 @@ fit_SE.combinations <- function(se,
 
       # average the top 10-percentile excess to get a single value 
       # for the excess
-      hsa_score[hsa_score$normalization_type == norm_type, "x"] <- ifelse(
+      scores[scores$normalization_type == norm_type, "hsa_score"] <- ifelse(
         is.null(h_excess), 
         NA, 
         mean(
-          h_excess$x[
-            h_excess$x >= 
-              stats::quantile(h_excess$x, 0.9, na.rm = TRUE)
+          h_excess$hsa_excess[
+            h_excess$hsa_excess >= 
+              stats::quantile(h_excess$hsa_excess, 0.9, na.rm = TRUE)
           ], 
           na.rm = TRUE
         )
       )
-      bliss_score[bliss_score$normalization_type == norm_type, "x"] <- ifelse(
-        is.null(b_excess), 
+      scores[scores$normalization_type == norm_type, "bliss_score"] <- ifelse(
+        is.null(bliss_excess), 
         NA, 
         mean(
-          b_excess$x[
-            b_excess$x >= 
-              stats::quantile(b_excess$x, 0.9, na.rm = TRUE)
+          bliss_excess$bliss_excess[
+            bliss_excess$bliss_excess >= 
+              stats::quantile(bliss_excess$bliss_excess, 0.9, na.rm = TRUE)
           ], 
           na.rm = TRUE
         )
       )
 
       if (all(vapply(isobologram_out, function(x) is.null(x) || all(is.na(x)), logical(1)))) {
-        CIScore_50[CIScore_50$normalization_type == norm_type, "x"] <-
-          CIScore_80[CIScore_80$normalization_type == norm_type, "x"] <- NA
+        scores[scores$normalization_type == norm_type, "CIScore_50"] <-
+          scores[scores$normalization_type == norm_type, "CIScore_80"] <- NA
       } else {
-        CIScore_50[CIScore_50$normalization_type == norm_type, "x"] <-
-          isobologram_out$df_all_AUC_log2CI$CI_100x[
+        scores[scores$normalization_type == norm_type, "CIScore_50"] <-
+          unique(isobologram_out$df_all_AUC_log2CI$CI_100x[
           isobologram_out$df_all_AUC_log2CI$iso_level ==
             min(isobologram_out$df_all_AUC_log2CI$iso_level[
               isobologram_out$df_all_AUC_log2CI$iso_level >= 0.5
-            ])]
-        CIScore_80[CIScore_80$normalization_type == norm_type, "x"] <-
-          isobologram_out$df_all_AUC_log2CI$CI_100x[
+            ])])
+        scores[scores$normalization_type == norm_type, "CIScore_80"] <-
+          unique(isobologram_out$df_all_AUC_log2CI$CI_100x[
           isobologram_out$df_all_AUC_log2CI$iso_level == 
             min(isobologram_out$df_all_AUC_log2CI$iso_level[
               isobologram_out$df_all_AUC_log2CI$iso_level >= 0.2
-            ])]
+            ])])
       }
       
-      b_excess$row_id <- av_matrix$row_id <- h_excess$row_id <- 
+      excess$row_id <-
         isobologram_out$df_all_iso_points$row_id <-
         isobologram_out$df_all_iso_curves$row_id <-
         col_fittings$row_id <-
-        hsa_score[, "row_id"] <- bliss_score[, "row_id"] <-
-        CIScore_50[, "row_id"] <-
-        CIScore_80[, "row_id"] <- metrics_merged$row_id <- i
-      b_excess$col_id <-
-        av_matrix$col_id <-
-        h_excess$col_id <- isobologram_out$df_all_iso_points$col_id <-
+        scores$row_id <-
+        metrics_merged$row_id <- i
+      excess$col_id <-
+        isobologram_out$df_all_iso_points$col_id <-
         isobologram_out$df_all_iso_curves$col_id <-
         col_fittings$col_id <-
-        hsa_score[, "col_id"] <- bliss_score[, "col_id"] <-
-        CIScore_50[, "col_id"] <-
-        CIScore_80[, "col_id"] <- metrics_merged$col_id <- j
+        scores$col_id <-
+        metrics_merged$col_id <- j
       
-      b_excess$normalization_type <- h_excess$normalization_type <- 
-        isobologram_out$df_all_iso_points$normalization_type <-
-        isobologram_out$df_all_iso_curves$normalization_type <- norm_type
+      isobologram_out$df_all_iso_points$normalization_type <-
+      isobologram_out$df_all_iso_curves$normalization_type <- norm_type
       
-      hsa_excess <- data.table::rbindlist(list(hsa_excess,
-                                               h_excess), fill = TRUE)
-      bliss_excess <- data.table::rbindlist(list(bliss_excess,
-                                                 b_excess), fill = TRUE)
       # check if it does not contain only ids
-      
-      if (!is.null(smooth_mx) && ncol(smooth_mx) != 2) {
-        smooth_mx <- data.table::rbindlist(list(smooth_mx, av_matrix), fill = TRUE)
-      } else {
-        smooth_mx <- av_matrix
-      }
-    
       if (is.null(all_iso_points)) {
         all_iso_points <- isobologram_out$df_all_iso_points
       } else {
@@ -398,50 +375,31 @@ fit_SE.combinations <- function(se,
       }
       
       metrics <- data.table::rbindlist(list(metrics, metrics_merged), fill = TRUE)
+      excess_full <- data.table::rbindlist(list(excess_full, excess), fill = TRUE)
     }
-    list(bliss_excess = bliss_excess,
-         hsa_excess = hsa_excess,
-         metrics = metrics,
+    list(metrics = metrics,
          all_iso_points = all_iso_points,
          isobolograms = isobolograms,
-         smooth_mx = smooth_mx,
-         bliss_score = bliss_score,
-         hsa_score = hsa_score,
-         CIScore_50 = CIScore_50,
-         CIScore_80 = CIScore_80)
+         excess = excess_full,
+         scores = scores)
   })
 
-  all_smooth_mx <- rbindParallelList(out, "smooth_mx")
-  all_hsa_excess <- rbindParallelList(out, "hsa_excess")
-  all_b_excess <- rbindParallelList(out, "bliss_excess")
+  excess <- rbindParallelList(out, "excess")
   all_iso_points <- rbindParallelList(out, "all_iso_points")
   all_isobolograms <- rbindParallelList(out, "isobolograms")
   all_metrics <- rbindParallelList(out, "metrics")
+  scores <- rbindParallelList(out, "scores")
+
   
-  bliss_score <- rbindParallelList(out, "bliss_score")
-  hsa_score <- rbindParallelList(out, "hsa_score")
-  CIScore_50 <- rbindParallelList(out, "CIScore_50")
-  CIScore_80 <- rbindParallelList(out, "CIScore_80")
-  
-  SummarizedExperiment::assays(se)[["SmoothMatrix"]] <- 
-    convertDFtoBumpyMatrixUsingIds(all_smooth_mx)
-  SummarizedExperiment::assays(se)[["BlissExcess"]] <- 
-    convertDFtoBumpyMatrixUsingIds(all_b_excess)
-  SummarizedExperiment::assays(se)[["HSAExcess"]] <- 
-    convertDFtoBumpyMatrixUsingIds(all_hsa_excess)
+  SummarizedExperiment::assays(se)[["excess"]] <- 
+    convertDFtoBumpyMatrixUsingIds(excess)
   SummarizedExperiment::assays(se)[["all_iso_points"]] <- 
     convertDFtoBumpyMatrixUsingIds(all_iso_points)
   SummarizedExperiment::assays(se)[["isobolograms"]] <- 
     convertDFtoBumpyMatrixUsingIds(all_isobolograms)
 
-  SummarizedExperiment::assays(se)[["BlissScore"]] <- 
-    convertDFtoBumpyMatrixUsingIds(bliss_score)
-  SummarizedExperiment::assays(se)[["HSAScore"]] <- 
-    convertDFtoBumpyMatrixUsingIds(hsa_score)
-  SummarizedExperiment::assays(se)[["CIScore_50"]] <- 
-    convertDFtoBumpyMatrixUsingIds(CIScore_50)
-  SummarizedExperiment::assays(se)[["CIScore_80"]] <- 
-    convertDFtoBumpyMatrixUsingIds(CIScore_80)
+  SummarizedExperiment::assays(se)[["scores"]] <- 
+    convertDFtoBumpyMatrixUsingIds(scores)
 
   SummarizedExperiment::assays(se)[[metrics_assay]] <- 
     convertDFtoBumpyMatrixUsingIds(all_metrics)