+++ R/packages.R
@@ -45,7 +45,8 @@ if (getRversion() >= "2.15.1") {
       "Duration",
       "isDay0",
       "record_id",
-      "ratio"
+      "ratio",
+      "smooth"
     ), 
     utils::packageName())
 }