+++ R/utils.R
@@ -208,9 +208,8 @@ data_model.data.table <- function(x) {
 get_data_type_to_data_model_mapping <- function() {
   c(
     `single-agent` = "single-agent",
-    "cotreatment" = "single-agent",
     "co-dilution" = "single-agent",
-    "matrix" = "combination"
+    "combination" = "combination"
   )
 }